requirements.txt
streamlit
matplotlib
pandas
numpy
seaborn
xlsxwriter
plotly
supabase
Flask
python-dotenv
requests
Flask-CORS
